+++ src/Serilog.Sinks.Http/Sinks/Http/Private/NonDurable/LogEventQueue.cs
@@ -20,16 +20,20 @@ namespace Serilog.Sinks.Http.Private.NonDurable
     public class LogEventQueue
     {
         private readonly Queue<string> queue;
-        private readonly int? queueLimit;
+        private readonly long? queueLimitBytes;
         private readonly object syncRoot = new();
 
-        public LogEventQueue(int? queueLimit = null)
+        private long queueBytes;
+
+        public LogEventQueue(long? queueLimitBytes = null)
         {
-            if (queueLimit < 1)
-                throw new ArgumentException("queueLimit must be either null or greater than 0", nameof(queueLimit));
+            if (queueLimitBytes < 1)
+                throw new ArgumentException("queueLimitBytes must be either null or greater than 0", nameof(queueLimitBytes));
 
             queue = new Queue<string>();
-            this.queueLimit = queueLimit;
+            this.queueLimitBytes = queueLimitBytes;
+
+            queueBytes = 0;
         }
 
         public void Enqueue(string logEvent)
@@ -45,11 +49,13 @@ namespace Serilog.Sinks.Http.Private.NonDurable
         {
             lock (syncRoot)
             {
-                if (queue.Count == queueLimit)
+                var logEventByteSize = ByteSize.From(logEvent);
+                if (queueBytes + logEventByteSize > queueLimitBytes)
                 {
                     return EnqueueResult.QueueFull;
                 }
 
+                queueBytes += logEventByteSize;
                 queue.Enqueue(logEvent);
                 return EnqueueResult.Ok;
             }
@@ -66,13 +72,15 @@ namespace Serilog.Sinks.Http.Private.NonDurable
                 }
 
                 logEvent = queue.Peek();
+                var logEventByteSize = ByteSize.From(logEvent);
 
-                if (ByteSize.From(logEvent) > logEventMaxSize)
+                if (logEventByteSize > logEventMaxSize)
                 {
                     logEvent = string.Empty;
                     return DequeueResult.MaxSizeViolation;
                 }
 
+                queueBytes -= logEventByteSize;
                 queue.Dequeue();
                 return DequeueResult.Ok;
             }